+++ core-ui/src/components/Predefined/Details/Sidecars/IstioListeners.js
@@ -4,9 +4,8 @@ import { useTranslation } from 'react-i18next';
 import { EMPTY_TEXT_PLACEHOLDER } from 'react-shared';
 import { LayoutPanel } from 'fundamental-react';
 import { LayoutPanelRow } from 'shared/components/LayoutPanelRow/LayoutPanelRow';
-import { WorkloadSelector } from 'shared/components/WorkloadSelector/WorkloadSelector';
 
-const IstioListeners = sidecar => {
+export const IstioListeners = sidecar => {
   const { t } = useTranslation();
   return (
     <>
@@ -25,35 +24,20 @@ const IstioListeners = sidecar => {
   );
 };
 
-const Hosts = ({ hosts }) => {
+const ListenerView = ({ listener, type, isEgress }) => {
   return (
-    <ul>
-      {hosts?.map((host, i) => (
-        <li key={i}>{host}</li>
-      ))}
-    </ul>
-  );
-};
-
-const Port = ({ port }) => {
-  const { t } = useTranslation();
-
-  return (
-    <div>
-      <LayoutPanelRow
-        name={t('sidecars.headers.port.number')}
-        value={port.number}
-      />
-      <LayoutPanelRow
-        name={t('sidecars.headers.port.protocol')}
-        value={port.protocol}
-      />
-      <LayoutPanelRow name={t('common.headers.name')} value={port.name} />
-      <LayoutPanelRow
-        name={t('sidecars.headers.port.target-point')}
-        value={port?.targetPoint || EMPTY_TEXT_PLACEHOLDER}
-      />
-    </div>
+    <LayoutPanel className="fd-margin--md">
+      <LayoutPanel.Header>
+        <LayoutPanel.Head title={type} />
+      </LayoutPanel.Header>
+      {listener?.map((properties, i) => (
+        <TrafficProperties
+          properties={properties}
+          key={i}
+          isEgress={isEgress}
+        />
+      )) || <LayoutPanel.Body>{EMPTY_TEXT_PLACEHOLDER}</LayoutPanel.Body>}
+    </LayoutPanel>
   );
 };
 
@@ -101,39 +85,34 @@ const TrafficProperties = ({ properties, isEgress }) => {
   );
 };
 
-const ListenerView = ({ listener, type, isEgress }) => {
+const Hosts = ({ hosts }) => {
   return (
-    <LayoutPanel className="fd-margin--md">
-      <LayoutPanel.Header>
-        <LayoutPanel.Head title={type} />
-      </LayoutPanel.Header>
-      {listener?.map((properties, i) => (
-        <TrafficProperties
-          properties={properties}
-          key={i}
-          isEgress={isEgress}
-        />
-      )) || <LayoutPanel.Body>{EMPTY_TEXT_PLACEHOLDER}</LayoutPanel.Body>}
-    </LayoutPanel>
+    <ul>
+      {hosts?.map((host, i) => (
+        <li key={i}>{host}</li>
+      ))}
+    </ul>
   );
 };
 
-export const SidecarsDetails = ({ DefaultRenderer, ...otherParams }) => {
+const Port = ({ port }) => {
   const { t } = useTranslation();
-  const customColumns = [
-    {
-      header: t('sidecars.headers.outbound-traffic-policy'),
-      value: sidecar =>
-        sidecar.spec?.outboundTrafficPolicy?.mode || EMPTY_TEXT_PLACEHOLDER,
-    },
-  ];
 
   return (
-    <DefaultRenderer
-      customComponents={[IstioListeners, WorkloadSelector]}
-      customColumns={customColumns}
-      singularName={t('persistent-volume-claims.name_singular')}
-      {...otherParams}
-    />
+    <div>
+      <LayoutPanelRow
+        name={t('sidecars.headers.port.number')}
+        value={port.number}
+      />
+      <LayoutPanelRow
+        name={t('sidecars.headers.port.protocol')}
+        value={port.protocol}
+      />
+      <LayoutPanelRow name={t('common.headers.name')} value={port.name} />
+      <LayoutPanelRow
+        name={t('sidecars.headers.port.target-point')}
+        value={port?.targetPoint || EMPTY_TEXT_PLACEHOLDER}
+      />
+    </div>
   );
 };
