+++ core-ui/src/components/Predefined/Details/Sidecars/Sidecars.details.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { EMPTY_TEXT_PLACEHOLDER } from 'react-shared';
+import { WorkloadSelector } from 'shared/components/WorkloadSelector/WorkloadSelector';
+import { IstioListeners } from './IstioListeners';
+
+export const SidecarsDetails = ({ DefaultRenderer, ...otherParams }) => {
+  const { t } = useTranslation();
+  const customColumns = [
+    {
+      header: t('sidecars.headers.outbound-traffic-policy'),
+      value: sidecar =>
+        sidecar.spec?.outboundTrafficPolicy?.mode || EMPTY_TEXT_PLACEHOLDER,
+    },
+  ];
+
+  return (
+    <DefaultRenderer
+      customComponents={[IstioListeners, WorkloadSelector]}
+      customColumns={customColumns}
+      singularName={t('persistent-volume-claims.name_singular')}
+      {...otherParams}
+    />
+  );
+};
